+++ CHANGELOG.md
@@ -13,6 +13,7 @@ For more detailed information about the changes see the history of the
 * refactored MO reordering of external QMPackages (#406)
 * Add defaults to XML (#319, #413)
 * fix build with gcc-10 (#422, #424)
+* Add name command line argument (#426)
 * fix usage of PROJECT_SOURCE_DIR in cmake (#431)
 * check that temperature is positive for KMC