+++ CHANGELOG.md
@@ -9,6 +9,9 @@ For more detailed information about the changes see the history of the
 * update boundary class (#457)
 * fix usage of PROJECT_SOURCE_DIR in cmake (#517)
 
+## Version 1.6.1 (released XX.04.20)
+* add support for cma-3 (#158) 
+
 ## Version 1.6 _SuperPelagia_ (released 17.04.20)
 * enforce periodic boundaries for dihedrals (#500)
 * add warning about dihedrals (#500)